requirements.txt
asgiref==3.6.0
Django==4.1.5
pytz==2022.7.1
sqlparse==0.4.3
mysqlclient==2.1.1
psycopg2-binary==2.9.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
python-decouple==3.7
whitenoise==6.3.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
Babel==2.11.0
requests==2.28.2
Pillow==9.4.0

